Created a separate class for each animal
