+++ src/Task7/Fish.java
@@ -1,0 +1,12 @@
+package Task7;
+
+    class Fish extends Animal {
+        public Fish (int id, String name, int age) {
+            super(id,name,age);
+        }
+        public void swim () {
+            System.out.println(name + " is swimming.");
+        }
+
+    }
+
