+++ src/Task7/Raven.java
@@ -1,0 +1,10 @@
+package Task7;
+
+class Raven extends Bird {
+
+    public Raven(int id, String name, int age, int maxFlightHeight) {
+        super(id, name, age, maxFlightHeight);
+    }
+
+}
+
